+++ src/Model/GalleryModelTrait.php
@@ -18,14 +18,6 @@ trait GalleryModelTrait
     }
 
     /**
-     * Get all users who have written an article
-     */
-    public static function getTest()
-    {
-        return self::query()->select('user_id', 'name', 'username')->groupBy('user_id', 'name')->join('@system_user', 'user_id = @system_user.id')->execute()->fetchAll();
-    }
-
-    /**
      * @Saving
      */
     public static function saving($event, Gallery $gallery)