+++ webpack.config.js
@@ -4,7 +4,7 @@ module.exports = [
         entry: {
             "gallery-edit": "./app/views/admin/gallery-edit",
             "gallery-index": "./app/views/admin/gallery-index",
-            "dashboard-gallery": "./app/components/dashboard-gallery.vue",
+            "gallery-dashboard": "./app/components/gallery-dashboard.vue",
             "settings": "./app/views/admin/settings",
             "link-gallery": "./app/components/link-gallery.vue",
             "gallery-meta": "./app/components/gallery-meta.vue",